server/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
ultralytics
torch==2.5.1+cpu
torchvision==0.20.1+cpu
tensorflow==2.15.0
pillow
requests
sqlalchemy
python-dotenv
flask-socketio
flask-cors
flask-jwt-extended
passlib
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
imutils
google-genai
